grupou-atividade-semana-1/final/outros/requirements.txt
pdfminer.six
nltk
matplotlib 
numpy